feat(stats): the customisable dashboard replaces the Stats page
- /stats now renders the dashboard (titled Reading Stats); /stats-lab
  redirects there and the sidebar Lab entries are gone
- old StatsPage.tsx deleted along with its duplicated chart code;
  orphaned Section component removed; dead OverviewHeadline dropped
- carried over from the old page: docs help link, BookAnimation loader,
  and the new-user onboarding empty state (shown only when there is
  genuinely no reading history, not just a quiet range)
- one-time discovery hint pointing at Edit, dismissed on use
- fix: 365-day heatmap bucketed days in UTC — evening sessions east of
  Greenwich lit up the wrong day; now uses local dates like every other
  chart
