+++ frontend/src/App.tsx
@@ -15,7 +15,6 @@ import { BookDetailPage } from '@/pages/BookDetailPage'
 import { AdminPage } from '@/pages/AdminPage'
 import { SettingsPage } from '@/pages/SettingsPage'
 import { StatsPage } from '@/pages/StatsPage'
-import { StatsLabPage } from '@/pages/StatsLabPage'
 import { BinderyPage } from '@/pages/BinderyPage'
 import { WishlistPage } from '@/pages/WishlistPage'
 import { api } from '@/lib/api'
@@ -120,14 +119,7 @@ function AppRoutes() {
             </ProtectedRoute>
           }
         />
-        <Route
-          path="/stats-lab"
-          element={
-            <ProtectedRoute>
-              <StatsLabPage />
-            </ProtectedRoute>
-          }
-        />
+        <Route path="/stats-lab" element={<Navigate to="/stats" replace />} />
         <Route
           path="/reader/:bookId"
           element={
